+++ src/test/java/com/doomedcat17/scpier/page/html/document/revision/LastRevisionDateScraperTest.java
@@ -1,7 +1,7 @@
 package com.doomedcat17.scpier.page.html.document.revision;
 
+import com.doomedcat17.scpier.TestDataProvider;
 import com.doomedcat17.scpier.exception.page.html.document.revision.RevisionDateException;
-import com.doomedcat17.scpier.testbox.TestDataProvider;
 import org.jsoup.nodes.Element;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
