+++ src/test/java/com/doomedcat17/scpier/scrappers/AudioScrapperTest.java
@@ -1,10 +1,10 @@
 package com.doomedcat17.scpier.scrappers;
 
+import com.doomedcat17.scpier.TestDataProvider;
 import com.doomedcat17.scpier.data.content.ContentNode;
 import com.doomedcat17.scpier.data.content.ContentNodeType;
 import com.doomedcat17.scpier.data.content.EmbedNode;
 import com.doomedcat17.scpier.scraper.audio.AudioScraper;
-import com.doomedcat17.scpier.testbox.TestDataProvider;
 import org.jsoup.nodes.Element;
 import org.junit.jupiter.api.Test;
