+++ src/test/java/com/doomedcat17/scpier/scrappers/LineScrapperTest.java
@@ -1,11 +1,11 @@
 package com.doomedcat17.scpier.scrappers;
 
+import com.doomedcat17.scpier.TestDataProvider;
 import com.doomedcat17.scpier.data.content.ContentNode;
 import com.doomedcat17.scpier.data.content.ParagraphNode;
 import com.doomedcat17.scpier.data.content.TextNode;
 import com.doomedcat17.scpier.scraper.line.LineScraper;
 import com.doomedcat17.scpier.scraper.line.TextNodeSplitter;
-import com.doomedcat17.scpier.testbox.TestDataProvider;
 import org.jsoup.nodes.Element;
 import org.junit.jupiter.api.Test;
