+++ src/test/java/com/doomedcat17/scpier/scrappers/StyleScrapperTest.java
@@ -1,7 +1,7 @@
 package com.doomedcat17.scpier.scrappers;
 
+import com.doomedcat17.scpier.TestDataProvider;
 import com.doomedcat17.scpier.scraper.text.StyleScraper;
-import com.doomedcat17.scpier.testbox.TestDataProvider;
 import org.jsoup.nodes.Element;
 import org.junit.jupiter.api.Test;
