+++ src/test/java/com/doomedcat17/scpier/scrappers/TextScraperTest.java
@@ -1,9 +1,9 @@
 package com.doomedcat17.scpier.scrappers;
 
+import com.doomedcat17.scpier.TestDataProvider;
 import com.doomedcat17.scpier.data.content.HyperlinkNode;
 import com.doomedcat17.scpier.data.content.TextNode;
 import com.doomedcat17.scpier.scraper.text.TextScraper;
-import com.doomedcat17.scpier.testbox.TestDataProvider;
 import org.jsoup.nodes.Element;
 import org.junit.jupiter.api.Test;
